refactor: user manager dir
