+++ frontend/rust-lib/flowy-core/src/deps_resolve/database_deps.rs
@@ -7,7 +7,7 @@ use collab_integrate::CollabKVDB;
 use flowy_database2::{DatabaseManager, DatabaseUser};
 use flowy_database_pub::cloud::DatabaseCloudService;
 use flowy_error::FlowyError;
-use flowy_user::manager::UserManager;
+use flowy_user::user_manager::UserManager;
 use lib_infra::priority_task::TaskDispatcher;
 
 pub struct DatabaseDepsResolver();
