+++ frontend/rust-lib/flowy-user-pub/src/lib.rs
@@ -1,3 +1,4 @@
 pub mod cloud;
 pub mod entities;
+
 pub const DEFAULT_USER_NAME: fn() -> String = || "Me".to_string();
