+++ frontend/rust-lib/flowy-user/src/lib.rs
@@ -5,11 +5,11 @@ mod anon_user;
 pub mod entities;
 mod event_handler;
 pub mod event_map;
-pub mod manager;
 mod migrations;
 mod notification;
 pub mod protobuf;
 pub mod services;
+pub mod user_manager;
 
 pub mod errors {
   pub use flowy_error::*;
