+++ frontend/rust-lib/flowy-user/src/services/db.rs
@@ -18,8 +18,8 @@ use flowy_user_pub::entities::{UserProfile, UserWorkspace};
 use lib_dispatch::prelude::af_spawn;
 use lib_infra::file_util::{unzip_and_replace, zip_folder};
 
-use crate::services::user_sql::UserTable;
-use crate::services::workspace_sql::UserWorkspaceTable;
+use crate::services::sqlite_sql::user_sql::UserTable;
+use crate::services::sqlite_sql::workspace_sql::UserWorkspaceTable;
 
 pub trait UserDBPath: Send + Sync + 'static {
   fn sqlite_db_path(&self, uid: i64) -> PathBuf;
