+++ frontend/rust-lib/flowy-user/src/services/sqlite_sql/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod user_sql;
+pub(crate) mod workspace_sql;
