+++ frontend/rust-lib/flowy-user/src/user_manager/mod.rs
@@ -1,0 +1,8 @@
+mod manager;
+pub(crate) mod manager_history_user;
+pub(crate) mod manager_user_awareness;
+pub(crate) mod manager_user_encryption;
+pub(crate) mod manager_user_workspace;
+mod user_login_state;
+
+pub use manager::*;
