+++ frontend/rust-lib/flowy-user/src/user_manager/user_login_state.rs
@@ -1,0 +1,11 @@
+use crate::migrations::MigrationUser;
+use flowy_user_pub::entities::{AuthResponse, Authenticator, UserProfile};
+
+/// recording the intermediate state of the sign-in/sign-up process
+#[derive(Clone)]
+pub struct UserAuthProcess {
+  pub user_profile: UserProfile,
+  pub response: AuthResponse,
+  pub authenticator: Authenticator,
+  pub migration_user: Option<MigrationUser>,
+}
